Working on scripts to send/receive 1 16 bit frame at a time for easier testing

+++ Transmit.py
@@ -72,7 +72,7 @@ QAM_values = data.modulate(data_bits, QAM, frame_length_bits)
 
 print(QAM_values)
 
-transmit = np.zeros(transmit_frames * frame_length_samples)
+transmit = np.zeros(transmit_frames * frame_length_samples, dtype = np.float32)
 
 if Modulation_type_OFDM:
 	print("Starting OFDM")
